+++ R/load-prepare_tag_sheet.R
@@ -35,25 +35,32 @@ prepare_tag_sheet <- function(path, header_line = 5, sheet_name = 2) {
   TAG_CODE_SPACE <- TAG_ID_CODE <- EST_TAG_LIFE <- UTC_RELEASE_DATE_TIME <-
     SEX <- RELEASE_LATITUDE <- RELEASE_LONGITUDE <- SCIENTIFIC_NAME <- NULL
 
-  tag_sheet <- readxl::read_excel(path, sheet = sheet_name, skip = header_line - 1)
-  tag_sheet <- tag_sheet %>% dplyr::mutate(
-    transmitter_id = paste(TAG_CODE_SPACE, TAG_ID_CODE, sep = "-"),
-    est_tag_life = convert_life_to_days(EST_TAG_LIFE)
+  tag_sheet <- readxl::read_excel(
+    path,
+    sheet = sheet_name,
+    skip = header_line - 1
   )
-  tag_sheet <- tag_sheet %>% dplyr::rename(
-    animal_id = "ANIMAL_ID   (floy tag ID, pit tag code, etc.)",
-    time = UTC_RELEASE_DATE_TIME,
-    sex = SEX,
-    latitude = RELEASE_LATITUDE,
-    longitude = RELEASE_LONGITUDE,
-    sci_name = SCIENTIFIC_NAME
-  )
+  tag_sheet <- tag_sheet %>%
+    dplyr::mutate(
+      transmitter_id = paste(TAG_CODE_SPACE, TAG_ID_CODE, sep = "-"),
+      est_tag_life = convert_life_to_days(EST_TAG_LIFE)
+    )
+  tag_sheet <- tag_sheet %>%
+    dplyr::rename(
+      animal_id = "ANIMAL_ID   (floy tag ID, pit tag code, etc.)",
+      time = UTC_RELEASE_DATE_TIME,
+      sex = SEX,
+      latitude = RELEASE_LATITUDE,
+      longitude = RELEASE_LONGITUDE,
+      sci_name = SCIENTIFIC_NAME
+    )
   return(tag_sheet)
 }
 
 # For converting tag life column to a count of days
 convert_life_to_days <- function(tagLife) {
-  sapply(tagLife,
+  sapply(
+    tagLife,
     function(x) {
       if (!grepl("[a-zA-Z]", x)) {
         return(as.integer(x))