+++ R/util-intern_pkg_man.r
@@ -8,7 +8,10 @@ add_internal_data <- function(x, rda_file) {
     load(rda_file, e) # load existing internals to new envir
     assign(deparse(substitute(x)), x, envir = e) # add/replace with new object
     save(
-      list = ls(e), file = rda_file, envir = e, compress = "bzip2",
+      list = ls(e),
+      file = rda_file,
+      envir = e,
+      compress = "bzip2",
       version = 2
     )
   } else {
@@ -23,7 +26,10 @@ remove_internal_data <- function(x, rda_file) {
   load(rda_file, e) # load existing internals to new envir
   rm(list = deparse(substitute(x)), envir = e) # remove unwanted object
   save(
-    list = ls(e), file = rda_file, envir = e, compress = "bzip2",
+    list = ls(e),
+    file = rda_file,
+    envir = e,
+    compress = "bzip2",
     version = 2
   )
 }