+++ R/util-vdat.r
@@ -1,6 +1,6 @@
 #' Convert an Innovasea VRL or VDAT file to a Fathom CSV file
 #'
-#' Use Innovasea's VDAT command line program VDAT.exe (distributed with Fathom
+#' Use Innovasea's VDAT command line program \code{vdat.exe} (distributed with Fathom
 #' Connect software) to make a CSV file containing data from a VRL or VDAT file
 #' in Fathom CSV format.
 #'
@@ -32,8 +32,8 @@
 #'  'Data Types to Include', 'Data Filter', 'Filename Suffix', 'Time Offset in
 #'  Hours', 'Split CSV by UTC Day'.)
 #'
-#' @param vdat_exe_path The full path to \code{VDAT.exe}. If \code{NULL}
-#'  (default) then the path to VDAT.exe must be in the PATH environment variable
+#' @param vdat_exe_path The full path to \code{vdat.exe}. If \code{NULL}
+#'  (default) then the path to vdat.exe must be in the PATH environment variable
 #'  of the system. See \code{\link{check_vdat}}.
 #'
 #' @param skip_pattern A regular expression used to exclude files from
@@ -51,9 +51,9 @@
 #'  CSV. Otherwise, only those files specified in \code{src} will be converted.
 #'
 #' @details Conversion is done by system call to the Innovasea program
-#'  \code{VDAT.exe} (included with Innovasea's Fathom Connect software;
+#'  \code{vdat.exe} (included with Innovasea's Fathom Connect software;
 #'  available at \url{https://support.fishtracking.innovasea.com/s/downloads}).
-#'  VDAT.exe must be available at the location specified by \code{vdat_exe_path}
+#'  vdat.exe must be available at the location specified by \code{vdat_exe_path}
 #'  or via system PATH environment variable. See also
 #'  \code{\link{check_vdat}}.
 #'
@@ -88,12 +88,14 @@
 #' # Check vdat.exe
 #' check_vdat()
 #'
-#' # all examples below assume path to VDAT.exe is in system PATH environment
+#' # all examples below assume path to vdat.exe is in system PATH environment
 #' # variable. If not (you get an error), add input argument 'vdat_exe_path'
-#' # with path directory with VDAT.exe.
+#' # with path directory with vdat.exe.
 #' # e.g.,
-#' # vdat_convert(vrl_files,
-#' #             vdat_exe_path = "C:/Program Files/Innovasea/Fathom")
+#' # vdat_convert(
+#' #   vrl_files,
+#' #   vdat_exe_path = "C:/Program Files/Innovasea/Fathom Connect"
+#' # )
 #'
 #' # get path to example VRL files in glatos
 #' vrl_files <- system.file("extdata", "detection_files_raw",
@@ -114,7 +116,7 @@
 #' # uncomment to open in file browser
 #' # utils::browseURL(temp_dir)
 #'
-#' # call VDAT.exe; default args
+#' # call vdatT.exe; default args
 #' vdat_convert(vrl_files2)
 #'
 #' # run again and overwrite
@@ -171,16 +173,17 @@
 #' }
 #'
 #' @export
-vdat_convert <- function(src,
-                         out_dir = NULL,
-                         output_format = "csv.fathom",
-                         overwrite = FALSE,
-                         recursive = FALSE,
-                         vdat_exe_path = NULL,
-                         skip_pattern = "-RLD_",
-                         show_progress = TRUE,
-                         diagn = FALSE,
-                         export_settings = NULL) {
+vdat_convert <- function(
+    src,
+    out_dir = NULL,
+    output_format = "csv.fathom",
+    overwrite = FALSE,
+    recursive = FALSE,
+    vdat_exe_path = NULL,
+    skip_pattern = "-RLD_",
+    show_progress = TRUE,
+    diagn = FALSE,
+    export_settings = NULL) {
   ##  Declare global variables for NSE & R CMD check
   src_dir <- src_file <- out_file <- out_file_exists <- src_to_convert <-
     written <- NULL
@@ -191,21 +194,23 @@ vdat_convert <- function(src,
   # Basic input checks
   stopifnot("Input 'src' must be character" = is.character(src))
   stopifnot(
-    "Input 'out_dir' must be character" =
-      is.character(out_dir) | is.null(out_dir)
+    "Input 'out_dir' must be character" = is.character(out_dir) |
+      is.null(out_dir)
   )
-  output_format <- match.arg(output_format,
+  output_format <- match.arg(
+    output_format,
     choices = c("csv.fathom", "csv.fathom.split")
   )
   stopifnot("Input 'overwrite' must be TRUE or FALSE" = is.logical(overwrite))
   stopifnot("Input 'recursive' must be TRUE or FALSE" = is.logical(recursive))
   stopifnot(
-    "Input 'vdat_exe_path' must be character or NULL" =
-      is.character(vdat_exe_path) | is.null(vdat_exe_path)
+    "Input 'vdat_exe_path' must be character or NULL" = is.character(
+      vdat_exe_path
+    ) |
+      is.null(vdat_exe_path)
   )
   stopifnot(
-    "Input 'show_progress' must be TRUE or FALSE" =
-      is.logical(show_progress)
+    "Input 'show_progress' must be TRUE or FALSE" = is.logical(show_progress)
   )
   stopifnot("Input 'diagn' must be TRUE or FALSE" = is.logical(diagn))
 
@@ -221,9 +226,7 @@ vdat_convert <- function(src,
   if (any(src_exists == FALSE)) {
     stop(
       "Input 'src' not found: \n ",
-      paste(src[src_exists == FALSE],
-        collapse = "\n "
-      )
+      paste(src[src_exists == FALSE], collapse = "\n ")
     )
   }
 
@@ -231,15 +234,14 @@ vdat_convert <- function(src,
   #  if not dir, assume file since previous step confirms existence
   src_type <- ifelse(dir.exists(src), "dir", "file")
 
-
   # Stop if mix of dir and file
   if (all(c("file", "dir") %in% src_type)) {
-    stop("Input arg 'src' must contain one or more ",
+    stop(
+      "Input arg 'src' must contain one or more ",
       "files or directories, but not both.",
       call. = FALSE
     )
   }
-
 
   # Check or set output directory
   out_dir_in <- out_dir
@@ -254,8 +256,10 @@ vdat_convert <- function(src,
     #  must either be NULL (use source dir for each file)
     #  scalar (same for all input files),
     #  or same length at src (specific to each input file or folder)
-    if (!(length(out_dir) == 1 |
-      length(out_dir) == length(src))) {
+    if (
+      !(length(out_dir) == 1 |
+        length(out_dir) == length(src))
+    ) {
       stop("Input 'out_dir' must be NULL, length one, or same length at 'src'.")
     }
 
@@ -266,9 +270,7 @@ vdat_convert <- function(src,
     if (any(out_dir_exists == FALSE)) {
       stop(
         "'out_dir' not found: \n ",
-        paste(out_dir[out_dir_exists == FALSE],
-          collapse = "\n "
-        )
+        paste(out_dir[out_dir_exists == FALSE], collapse = "\n ")
       )
     }
 
@@ -295,7 +297,8 @@ vdat_convert <- function(src,
 
     ftp <- ftp[,
       list(
-        src_file = list.files(src_dir,
+        src_file = list.files(
+          src_dir,
           full.names = TRUE,
           recursive = recursive,
           pattern = paste(
@@ -316,9 +319,7 @@ vdat_convert <- function(src,
       warning(paste0(
         "No VRL or VDAT files were ",
         "found in 'src': \n ",
-        paste(src,
-          collapse = "\n "
-        )
+        paste(src, collapse = "\n ")
       ))
 
       return(NA_character_)
@@ -334,34 +335,36 @@ vdat_convert <- function(src,
   # File extension (and type) depends on output_format
   output_file_ext <-
     data.table::fcase(
-      output_format == "csv.fathom", "csv",
-      output_format == "csv.fathom.split", "csv-fathom-split"
+      output_format == "csv.fathom",
+      "csv",
+      output_format == "csv.fathom.split",
+      "csv-fathom-split"
     )
 
   # Set output path(s) and file name(s)
-  ftp[, out_file := file.path(
-    out_dir,
-    gsub(
-      paste(paste0("\\.", supported_ext, "$"),
-        collapse = "|"
-      ),
-      paste0(".", output_file_ext),
-      basename(src_file),
-      ignore.case = TRUE
+  ftp[
+    ,
+    out_file := file.path(
+      out_dir,
+      gsub(
+        paste(paste0("\\.", supported_ext, "$"), collapse = "|"),
+        paste0(".", output_file_ext),
+        basename(src_file),
+        ignore.case = TRUE
+      )
     )
-  )]
-
+  ]
 
   # Ignore existing files if overwrite is false
   ftp[, out_file_exists := (file.exists(out_file))]
   ftp[, src_to_convert := !out_file_exists | overwrite]
 
-
   # Convert files
 
   # Loop through files so that progress can be displayed
   message(
-    "Converting ", sum(ftp$src_to_convert),
+    "Converting ",
+    sum(ftp$src_to_convert),
     " VRL/VDAT file(s) to Fathom CSV..."
   )
 
@@ -390,9 +393,12 @@ vdat_convert <- function(src,
     vdat_call <- with(
       ftp,
       paste0(
-        "convert --format=", output_format, " \"",
+        "convert --format=",
+        output_format,
+        " \"",
         src_file[i],
-        "\" --timec=default --output \"", out_dir[i],
+        "\" --timec=default --output \"",
+        out_dir[i],
         "\""
       )
     )
@@ -408,25 +414,33 @@ vdat_convert <- function(src,
 
     # if warning; make new warning with error message
     if (!is.null(attr(msg_i, "status"))) {
-      ftp[i, `:=`(
-        vdat_error = msg_i[1],
-        vdat_status = attr(msg_i, "status")
-      )]
+      ftp[
+        i,
+        `:=`(
+          vdat_error = msg_i[1],
+          vdat_status = attr(msg_i, "status")
+        )
+      ]
 
       if (diagn) {
         warning(
-          "\nError converting ", basename(ftp$src_file[i]), " :\n",
-          "\trunning command'", vdat_cmd, " ", vdat_call,
-          " had status ", ftp$vdat_status[i], ":\n",
+          "\nError converting ",
+          basename(ftp$src_file[i]),
+          " :\n",
+          "\trunning command'",
+          vdat_cmd,
+          " ",
+          vdat_call,
+          " had status ",
+          ftp$vdat_status[i],
+          ":\n",
           ftp$vdat_error[i]
         )
       }
     }
 
     if (show_progress) {
-      setTxtProgressBar(pb,
-        value = sum(ftp$src_to_convert[1:i])
-      )
+      setTxtProgressBar(pb, value = sum(ftp$src_to_convert[1:i]))
     }
   } # end i
 
@@ -440,7 +454,8 @@ vdat_convert <- function(src,
 
   message(
     "Done. ",
-    sum(ftp[src_to_convert == TRUE]$written, na.rm = TRUE), " of ",
+    sum(ftp[src_to_convert == TRUE]$written, na.rm = TRUE),
+    " of ",
     nrow(ftp),
     " file(s) written to disk."
   )
@@ -450,12 +465,13 @@ vdat_convert <- function(src,
     with(
       ftp,
       message(
-        "\n! ", sum(src_to_convert == FALSE),
+        "\n! ",
+        sum(src_to_convert == FALSE),
         " file(s) skipped (already exists &",
-        " 'overwrite' = ", overwrite, "):\n   ",
-        paste(basename(out_file[src_to_convert == FALSE]),
-          collapse = "\n   "
-        ),
+        " 'overwrite' = ",
+        overwrite,
+        "):\n   ",
+        paste(basename(out_file[src_to_convert == FALSE]), collapse = "\n   "),
         "\n"
       )
     )
@@ -466,13 +482,16 @@ vdat_convert <- function(src,
     with(
       ftp[written == FALSE],
       message(
-        "\n! ", sum(written == FALSE),
+        "\n! ",
+        sum(written == FALSE),
         " file(s) not written due to errors:\n   ",
         paste(
           paste0(
             basename(src_file),
-            ": (status ", vdat_status,
-            ") ", vdat_error
+            ": (status ",
+            vdat_status,
+            ") ",
+            vdat_error
           ),
           collapse = "\n   "
         ),
@@ -485,13 +504,13 @@ vdat_convert <- function(src,
 }
 
 
-#' Check path to Innovasea program VDAT.exe
+#' Check path to Innovasea program \code{vdat.exe}
 #'
-#' @param vdat_exe_path The full path to \code{VDAT.exe}. If \code{NULL}
-#'  (default) then the path to VDAT.exe must be in the PATH environment variable
-#'  of the system.
+#' @param vdat_exe_path The full path to \code{vdat.exe}. If \code{NULL}
+#'  (default) then the path to \code{vdat.exe} must be in the PATH environment
+#'  variable of the system.
 #'
-#' @returns Character string with command for calling VDAT.exe via
+#' @returns Character string with command for calling \code{vdat.exe} via
 #'   \code{system2}'s \code{command} argument.
 #'
 #' @examples
@@ -501,12 +520,12 @@ vdat_convert <- function(src,
 #' check_vdat()
 #'
 #'
-#' # use path to directory containing VDAT.exe
-#' check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom")
+#' # use path to directory containing vdat.exe
+#' check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom Connect")
 #'
 #'
-#' # use full path to VDAT.exe
-#' check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom/VDAT.exe")
+#' # use full path to vdat.exe
+#' check_vdat(vdat_exe_path = "C:/Program Files/Innovasea/Fathom Connect/vdat.exe")
 #' }
 #'
 #' @export
@@ -515,36 +534,29 @@ check_vdat <- function(vdat_exe_path = NULL) {
     vdat_cmd <- "VDAT"
 
     if (Sys.which(vdat_cmd) == "") {
-      stop("VDAT.exe not found in system PATH ",
-        "variable.",
-        call. = FALSE
-      )
+      stop("vdat.exe not found in system PATH ", "variable.", call. = FALSE)
     }
   } else {
-    # remove VDAT.exe from vdat_exe_path if present
-    vdat_exe_dir <- ifelse(grepl("vdat.exe$",
-      vdat_exe_path,
-      ignore.case = TRUE
-    ),
-    dirname(vdat_exe_path),
-    vdat_exe_path
+    # remove vdat.exe from vdat_exe_path if present
+    vdat_exe_dir <- ifelse(
+      grepl("vdat\\.exe$", vdat_exe_path, ignore.case = TRUE),
+      dirname(vdat_exe_path),
+      vdat_exe_path
     )
 
-    vdat_exe_file <- file.path(vdat_exe_dir, "VDAT.exe")
+    vdat_exe_file <- file.path(vdat_exe_dir, "vdat.exe")
 
-    # Check path to VDAT.exe
+    # Check path to vdat.exe
     if (!file.exists(vdat_exe_file)) {
-      stop("VDAT.exe not found at specified ",
-        "path.",
-        call. = FALSE
-      )
+      stop("vdat.exe not found at specified ", "path.", call. = FALSE)
     }
 
     vdat_cmd <- vdat_exe_file
 
     # Check if path can be reached via system call
     if (Sys.which(vdat_cmd) == "") {
-      stop("VDAT.exe found but could not be ",
+      stop(
+        "vdat.exe found but could not be ",
         "reached via system call.",
         call. = FALSE
       )
@@ -555,26 +567,26 @@ check_vdat <- function(vdat_exe_path = NULL) {
 }
 
 
-#' Get version of local installation of Innovasea program VDAT.exe
+#' Get version of local installation of Innovasea program vdat.exe
 #'
-#' @param vdat_exe_path The full path to \code{VDAT.exe}. If \code{NULL}
-#'  (default) then the path to VDAT.exe must be in the PATH environment variable
-#'  of the system. See \code{\link{check_vdat}}.
+#' @param vdat_exe_path The full path to \code{vdat.exe}. If \code{NULL}
+#'  (default) then the path to \code{vdat.exe} must be in the PATH environment
+#'  variable of the system. See \code{\link{check_vdat}}.
 #'
 #' @returns
 #' A list with \code{version} (version number) and \code{long_version} (full
-#' string returned by VDAT.exe).
+#' string returned by vdat.exe).
 #'
 #' @examples
 #' \dontrun{
 #'
-#' # use if VDAT.exe in Windows system PATH variable
+#' # use if vdat.exe in Windows system PATH variable
 #' get_local_vdat_version()
 #'
-#' # or specify path to VDAT.exe
+#' # or specify path to vdat.exe
 #' get_local_vdat_version(
 #'   vdat_exe_path =
-#'     "C:/Program Files/Innovasea/Fathom/VDAT.exe"
+#'     "C:/Program Files/Innovasea/Fathom Connect/vdat.exe"
 #' )
 #' }
 #'
@@ -583,7 +595,7 @@ get_local_vdat_version <- function(vdat_exe_path = NULL) {
   # Check path to vdat.exe and get (valid) command arg for system2 call
   vdat_cmd <- check_vdat(vdat_exe_path)
 
-  # invoke VDAT.exe
+  # invoke vdat.exe
   vdat_call <- "--version"
 
   vdat_version <- system2(vdat_cmd, vdat_call, stdout = TRUE)
@@ -601,29 +613,29 @@ get_local_vdat_version <- function(vdat_exe_path = NULL) {
 }
 
 
-#' Get schema from local installation of Innovasea program VDAT.exe
+#' Get schema from local installation of Innovasea program \code{vdat.exe}
 #'
-#' @param vdat_exe_path The full path to \code{VDAT.exe}. If \code{NULL}
-#'  (default) then the path to VDAT.exe must be in the PATH environment variable
-#'  of the system. See \code{\link{check_vdat}}.
+#' @param vdat_exe_path The full path to \code{vdat.exe}. If \code{NULL}
+#'  (default) then the path to \code{vdat.exe} must be in the PATH environment
+#'  variable of the system. See \code{\link{check_vdat}}.
 #'
 #' @details A bug in vdat.exe version 9 (confirmed on vdat-9.3.0) will cause
 #'   this function to return an empty list. Fixed in vdat.exe version 10
 #'   (confirmed on vdat-10.6.0).
 #'
 #' @returns
-#' Schema (template) of VDAT CSV produced by installed version of VDAT.exe.
+#' Schema (template) of VDAT CSV produced by installed version of \code{vdat.exe}.
 #'
 #' @examples
 #' \dontrun{
 #'
-#' # use if VDAT.exe in Windows system PATH variable
+#' # use if vdat.exe in Windows system PATH variable
 #' get_local_vdat_template()
 #'
-#' # or specify path to VDAT.exe
+#' # or specify path to vdat.exe
 #' get_local_vdat_template(
 #'   vdat_exe_path =
-#'     "C:/Program Files/Innovasea/Fathom/VDAT.exe"
+#'     "C:/Program Files/Innovasea/Fathom Connect/vdat.exe"
 #' )
 #' }
 #'
@@ -632,7 +644,7 @@ get_local_vdat_template <- function(vdat_exe_path = NULL) {
   # Check path to vdat.exe and get (valid) command arg for system2 call
   vdat_cmd <- check_vdat(vdat_exe_path)
 
-  # Invoke VDAT.exe
+  # Invoke vdat.exe
   vdat_call <- "template --format=csv.fathom"
 
   vdat_schema <- system2(vdat_cmd, vdat_call, stdout = TRUE)
@@ -640,7 +652,6 @@ get_local_vdat_template <- function(vdat_exe_path = NULL) {
   # remove BOM from start of first row if present
   check_bom <- vdat_schema[1]
   Encoding(check_bom) <- "latin1"
-
 
   if (length(suppressMessages(tools::showNonASCII(check_bom))) > 0) {
     Encoding(vdat_schema[1]) <- "latin1"